refactor(server): move onCommands out to services
